rack/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.129.0
h11==0.16.0
idna==3.11
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.5.0
python-dotenv==1.2.1
python-multipart==0.0.22
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0


python-docx==1.1.2